requirements.txt
django==3.0.3
djangorestframework==3.14.0
django-cors-headers==3.11.0
psycopg2==2.8.4
djangorestframework-simplejwt==5.2.2
firebase-admin==6.1.0